Move tests from filters.py
